pkg/prometheus: move LabelConfigValidator to validation package
This move improves code organization and encapsulation by isolating
validation logic from resource selection. It also resolves issues
related to accessing private fields/methods across files.

The 'validate' method was renamed to 'ValidateRelabelConfig' to allow
it to be tested and used from outside the package.

Updated references in:
- pkg/prometheus/resource_selector.go
- pkg/prometheus/promcfg.go
- pkg/prometheus/server/operator.go
- pkg/prometheus/resource_selector_test.go

Fixes: #8209
